Refactor endpoint allocation code
The main change here is to allocate the IN and OUT endpoints for
endpoint 0 first, rather than as a final step in Driver::start().
Driver::start() does not have a way to return an error, so it's better
for us to account for EP0's FIFO space requirements up front.  This way
if we run out of FIFO space we can return an error when allocating other
endpoints, rather than having to panic in Driver::start().

This also restores the logic which allows for arbitrary mapping between
IN endpoints and TX FIFOs, although in practice this should still always
end up allocating TX FIFO N to IN endpoint N.